requirements.txt
python-dotenv~=0.21.0
motor~=3.1.1
pymongo~=4.3.3
aiosqlite~=0.18.0
fastapi~=0.88.0
Pillow~=9.3.0
pydantic~=1.10.2
aioboto3~=10.2.0
uvicorn~=0.20.0
aiohttp~=3.8.3
pytest~=7.2.0
qrcode~=7.4.2